Add support for wheel mode for xl-320 motors.
