+++ pypot/_version.py
@@ -1,1 +1,1 @@
-__version__ = '2.4.0'
+__version__ = '2.5.0'
